refactor: move Offset source files into src/Editor directory
