+++ src/CMakeLists.txt
@@ -21,8 +21,8 @@ target_sources(kilo
         "${PROJECT_SOURCE_DIR}/src/EditorConfig/EditorConfig.hpp"
         "${PROJECT_SOURCE_DIR}/src/EditorConfig/EditorConfig.cpp"
 
-        "${PROJECT_SOURCE_DIR}/src/Offset/Offset.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Offset/Offset.cpp"
+        "${PROJECT_SOURCE_DIR}/src/Editor/Offset.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Editor/Offset.cpp"
 
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.hpp"
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.cpp"
