+++ tests/Editor.test.cpp
@@ -24,7 +24,7 @@
 #include "Editor/Editor.hpp"
 
 #include "Cursor/Cursor.hpp"
-#include "Offset/Offset.hpp"
+#include "Editor/Offset.hpp"
 #include "ScreenBuffer/ScreenBuffer.hpp"
 #include "Utilities/Utilities.hpp"
 #include "Window/Window.hpp"
